+++ src/backend/marsha/core/tests/test_utils_medialive_utils.py
@@ -10,6 +10,9 @@ from ..factories import VideoFactory
 from ..utils import medialive_utils
 
 
+# pylint: disable=too-many-lines
+
+
 class MediaLiveUtilsTestCase(TestCase):
     """Test medialive utils."""
 
@@ -812,6 +815,47 @@ class MediaLiveUtilsTestCase(TestCase):
             )
 
             medialive_utils.delete_aws_element_stack(video)
+            medialive_stubber.assert_no_pending_responses()
+
+    @override_settings(AWS_MEDIALIVE_INPUT_WAITER_DELAY=0)
+    @override_settings(AWS_MEDIALIVE_INPUT_WAITER_MAX_ATTEMPTS=0)
+    def test_delete_aws_elemental_stack_waiter_attemps_exceeded(self):
+        """Medialive channel should be deleted and input waiter raised a WaiterError.
+        The function should not fail but sentry should capture the exception."""
+        video = VideoFactory(
+            live_info={
+                "medialive": {
+                    "input": {"id": "medialive_input1"},
+                    "channel": {"id": "medialive_channel1"},
+                },
+                "mediapackage": {
+                    "endpoints": {
+                        "hls": {"id": "mediapackage_endpoint1"},
+                    },
+                    "channel": {"id": "mediapackage_channel1"},
+                },
+            }
+        )
+
+        with Stubber(
+            medialive_utils.medialive_client
+        ) as medialive_stubber, mock.patch.object(
+            medialive_utils, "capture_exception"
+        ) as mock_capture_exception:
+            medialive_stubber.add_response(
+                "delete_channel",
+                expected_params={"ChannelId": "medialive_channel1"},
+                service_response={},
+            )
+            medialive_stubber.add_response(
+                "describe_input",
+                expected_params={"InputId": "medialive_input1"},
+                service_response={"State": "ATTACHED"},
+            )
+
+            medialive_utils.delete_aws_element_stack(video)
+            medialive_stubber.assert_no_pending_responses()
+            mock_capture_exception.assert_called_once()
 
     def test_wait_medialive_channel_is_created(self):
         """Should call describe_channel while state is not IDLE."""